First commit
A simple tool to download folder from MF. This is solely for learning purpose, I do not take any responsibility for wrong uses of the tool.

I only tested it in Windows, but the tool should run well on other platforms, with a little tweaks.

On windows, you need to have Firefox and Python3 installed. Also python libraries BeautifulSoup and Selenium.

Simply open the folderdownload.py in python console, paste the link and press enter. Please wait for a couple of minutes before it's done.

You will get a txt file, which can be used to batch download using a download manager, such as IDM.